[Android] Fixes for new lint warnings
There were added new lint checks in C131. This commit fixes the following warnings:
warning: [UnnecessaryParentheses] These grouping parentheses are unnecessary; it is unlikely the code will be misinterpreted without them
warning: [OperatorPrecedence] Use grouping parenthesis to make the operator precedence explicit
warning: [NarrowCalculation] This product of integers could overflow before being implicitly cast to a long
warning: [NarrowCalculation] This division will discard the fractional part of the result, despite being assigned to a float
warning: [ImmutableEnumChecker] enums should be immutable
warning: [Finalize] Do not override finalize
warning: [DoubleBraceInitialization] Prefer collection factory methods or builders to the double-brace initialization pattern
warning: [UnusedTypeParameter] This type parameter is unused and can be removed
warning: [UnnecessaryStringBuilder] Prefer string concatenation over explicitly using `StringBuilder#append`, since `+` reads better and has equivalent or better performance

+++ android/java/org/chromium/chrome/browser/qrreader/CameraSource.java
@@ -458,7 +458,7 @@ public class CameraSource {
             currentZoom = parameters.getZoom() + 1;
             float newZoom;
             if (scale > 1) {
-                newZoom = currentZoom + scale * (maxZoom / 10);
+                newZoom = currentZoom + scale * (maxZoom / 10f);
             } else {
                 newZoom = currentZoom * scale;
             }
@@ -764,7 +764,9 @@ public class CameraSource {
                     mFocusMode)) {
                 parameters.setFocusMode(mFocusMode);
             } else {
-                Log.i(TAG, "Camera focus mode: " + mFocusMode + " is not supported on this device.");
+                Log.i(
+                        TAG,
+                        "Camera focus mode: " + mFocusMode + " is not supported on this device.");
             }
         }
 
@@ -776,7 +778,9 @@ public class CameraSource {
                     mFlashMode)) {
                 parameters.setFlashMode(mFlashMode);
             } else {
-                Log.i(TAG, "Camera flash mode: " + mFlashMode + " is not supported on this device.");
+                Log.i(
+                        TAG,
+                        "Camera flash mode: " + mFlashMode + " is not supported on this device.");
             }
         }
 
@@ -1007,14 +1011,15 @@ public class CameraSource {
     }
 
     /**
-     * Creates one buffer for the camera preview callback.  The size of the buffer is based off of
+     * Creates one buffer for the camera preview callback. The size of the buffer is based off of
      * the camera preview size and the format of the camera image.
      *
      * @return a new preview buffer of the appropriate size for the current camera settings
      */
     private byte[] createPreviewBuffer(Size previewSize) {
         int bitsPerPixel = ImageFormat.getBitsPerPixel(ImageFormat.NV21);
-        long sizeInBits = previewSize.getHeight() * previewSize.getWidth() * bitsPerPixel;
+        long sizeInBits =
+                previewSize.getHeight() * (long) previewSize.getWidth() * (long) bitsPerPixel;
         int bufferSize = (int) Math.ceil(sizeInBits / 8.0d) + 1;
 
         //
